Decoupling
- Abstracted logic from MainActivity into DataHandler + interface and AuthorizationHandler + interface
- Moved BluetoothServer & SocketHandler into separate package
