+++ Phone/app/src/main/java/com/facilitation/phone/MainActivity.kt
@@ -10,26 +10,28 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.facilitation.phone.databinding.ActivityMainBinding
-import com.facilitation.phone.model.SpotifyPlaylist
-import com.facilitation.phone.model.TrackDTO
+import com.facilitation.phone.utility.authorization.AuthorizationHandler
+import com.facilitation.phone.utility.authorization.IAuthorization
+import com.facilitation.phone.utility.data.DataHandler
+import com.facilitation.phone.utility.data.IData
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.gson.Gson
 import com.spotify.sdk.android.auth.AuthorizationClient
-import com.spotify.sdk.android.auth.AuthorizationRequest
 import com.spotify.sdk.android.auth.AuthorizationResponse
-import okhttp3.OkHttpClient
-import okhttp3.Request
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
     private lateinit var navController: NavController
-    private val gson = Gson()
+    private lateinit var dataHandler: IData
+    private var authorizationHandler: IAuthorization = AuthorizationHandler()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        dataHandler = DataHandler(this)
 
         val navView: BottomNavigationView = binding.navView
 
@@ -44,15 +46,7 @@ class MainActivity : AppCompatActivity() {
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
 
-        authorizationSpotify()
-    }
-
-    private fun authorizationSpotify() {
-        val builder: AuthorizationRequest.Builder =
-            AuthorizationRequest.Builder(getString(R.string.client_id), AuthorizationResponse.Type.TOKEN, getString(R.string.redirect_uri))
-        builder.setScopes(arrayOf("streaming"))
-        val request: AuthorizationRequest = builder.build()
-        AuthorizationClient.openLoginActivity(this, 9485, request)
+        authorizationHandler.authorizationSpotify(this)
     }
 
     @Deprecated("Deprecated in Java")
@@ -85,38 +79,7 @@ class MainActivity : AppCompatActivity() {
 
     private fun retrievePlaylistTracks() {
         Thread {
-        val sharedPreferencesSpotify = getSharedPreferences("SPOTIFY", 0)
-        val token = sharedPreferencesSpotify.getString("token", null)
-        val client = OkHttpClient()
-        val request = Request.Builder()
-            .url(getString(R.string.playlist_uri))
-            .header("Authorization", "Bearer $token")
-            .build()
-
-        val response = client.newCall(request).execute()
-        val responseData = response.body?.string()
-        deserializeIntoTracks(responseData)
+            dataHandler.retrievePlaylist()
         }.start()
-    }
-
-    private fun deserializeIntoTracks(response: String?) {
-        val playlist = gson.fromJson(response, SpotifyPlaylist::class.java)
-
-        //Constructs tracksDTO list based on the playlist retrieved from Spotify
-        val tracksDTO: List<TrackDTO> = playlist.tracks.items.map { item ->
-            //In case a song has multiple artists, this will take care of it
-            val concatenatedArtists = item.track.artists.joinToString(", ") { it.name }
-            TrackDTO(item.track.name, concatenatedArtists, item.track.uri, false)
-        }
-        storeTracksDTO(tracksDTO)
-    }
-
-    private fun storeTracksDTO(tracksDTO: List<TrackDTO>) {
-        val tracksDTOJson = gson.toJson(tracksDTO)
-        val sharedPreferences = getSharedPreferences("SPOTIFY", 0)
-        val editor = sharedPreferences.edit()
-        editor.putString("tracksDTOJson", tracksDTOJson)
-        editor.apply()
-        Log.i("VuzixSidekick", "Playlist with ${tracksDTO.size} tracks successfully retrieved and saved in shared preferences")
     }
 }
