+++ Phone/app/src/main/java/com/facilitation/phone/utility/authorization/IAuthorization.kt
@@ -1,0 +1,7 @@
+package com.facilitation.phone.utility.authorization
+
+import android.app.Activity
+
+interface IAuthorization {
+    fun authorizationSpotify(activity: Activity)
+}
