+++ Phone/app/src/main/java/com/facilitation/phone/utility/data/IData.kt
@@ -1,0 +1,5 @@
+package com.facilitation.phone.utility.data
+
+interface IData {
+    fun retrievePlaylist()
+}
