+++ Phone/app/src/main/java/com/facilitation/phone/viewModel/HomeViewModel.kt
@@ -2,9 +2,9 @@ package com.facilitation.phone.viewModel
 
 import android.app.Activity
 import android.app.Application
-import androidx.lifecycle.AndroidViewModel
 import android.util.Log
-import com.facilitation.phone.utility.BluetoothServer
+import androidx.lifecycle.AndroidViewModel
+import com.facilitation.phone.utility.communication.BluetoothServer
 
 
 class HomeViewModel(private val appContext: Application) : AndroidViewModel(appContext) {
@@ -14,7 +14,7 @@ class HomeViewModel(private val appContext: Application) : AndroidViewModel(appC
         if(!serverRunning) {
             serverRunning = BluetoothServer(appContext, activity).startServer()
         }
-        else{
+        else {
             Log.i("VuzixSidekick", "Bluetooth server is already running")
         }
     }
